requirements.txt
numpy==1.13.1
pandas==0.20.3
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
websocket-client==0.44.0